+++ age-core/CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to Rust's notion of
 to 1.0.0 are beta releases.
 
 ## [Unreleased]
+
+## [0.9.0] - 2022-10-27
 ### Changed
+- MSRV is now 1.59.0.
 - Migrated to `aead 0.5`.
 
 ## [0.8.0] - 2022-05-02